feat(MetaDataWidget): add graph view and term depiction tabs
The graphViewWidget and TermDepictionWidget were missing from the
metadata widget as they are relevant to the selected term. This commit
also changes the term in the default story for the metadata widget to a term in
Uberon ontology to have a working example for the term depiction tab.
The default for the new tabs are true so current users do not have to
change anything. They can set to false if they want to hide them.